backend/core/email_validate.py
import re
import dns.resolver
from disposable_email_domains import blocklist


def is_valid_email_username(username: str) -> str:
    """Can it be the username@domain part of an email address?"""
    pattern = r"^[a-zA-Z0-9.\-_\+]+$"
    if re.match(pattern, username):
        return True
    return False


def is_disposable_domain(hostname: str) -> bool:
    """Check if the domain is from a disposable email provider like mailinator.com"""
    if hostname in blocklist:
        return True
    return False


def hostname_has_mx(hostname: str) -> bool:
    try:
        records = dns.resolver.resolve(hostname, "MX")
    except (dns.resolver.NoAnswer, dns.resolver.NXDOMAIN, dns.resolver.LifetimeTimeout):
        return False
    return True


def validate_email(
    email: str,
    check_dns: bool = True,
    check_disposable_domains: bool = True,
):
    """Validate an email address.

    Returns:
    (status, canonical_email, message)
        - status: boolean (True = email is ok),
        - canonical_email: str - proper email without whitespace or mixed case
        - message: str error explanation
    """
    if not isinstance(email, str) or "@" not in email:
        return (False, None, f"Invalid email {email} - no @ symbol")
    email = email.lower().rstrip().lstrip()

    # to let tests pass
    if email == "test@test.com":
        return (True, email, f"Email {email} is valid for testing only")

    user, hostname = email.split("@", 1)
    if not is_valid_email_username(user):
        return (False, None, f"Invalid username {user} in email {email}")

    if '.' not in hostname:
        return (False, None, f"Invalid domain {hostname} - must have at least one dot")

    for label in hostname.split('.'):
        if not label or len(label) > 63:
            return (False, None, f"Invalid domain {hostname} - each part must be between 1 and 63 characters")
        if not all(c.isalnum() or c == '-' for c in label):
            return (False, None, f"Invalid domain {hostname} - can only contain letters, numbers, and hyphens")
        if label.startswith('-') or label.endswith('-'):
            return (False, None, f"Invalid domain {hostname} - parts cannot start or end with hyphens")

    if check_dns and not hostname_has_mx(hostname):
        return (False, None, f"Domain {hostname} does not exist "
                "or is not configured to receive email messages")

    if check_disposable_domains and is_disposable_domain(hostname):
        return (False, None, f"Email {email} is a disposable email address")

    return (True, email, f"Email {email} is valid")
