backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
python-jose[cryptography]>=3.3.0
google-cloud-firestore>=2.19.0
httpx>=0.27.0
mailersend>=0.6.0
dnspython>=2.7.0
disposable-email-domains>=0.0.103
python-multipart>=0.0.18
